pyproject.toml
[project]
name = "ai-poc"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "ag-ui-protocol>=0.1.14",
    "fastapi[all]>=0.135.2",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "langgraph-checkpoint-redis>=0.4.0",
    "openpyxl>=3.1.5",
    "pymongo>=4.16.0",
    "mem0ai>=0.1.0",
    "motor>=3.6.0",
    "langchain-memgraph>=0.1.0",
    "rank-bm25>=0.2.2",
    "aiohttp>=3.13.5",
    "multidict>=6.7.1",
]
